requirements.txt
streamlit
opencv-python 
transformers 
pillow 
pyttsx3 
pytesseract 
scikit-learn
numpy
pytesseract